+++ src/main/java/kitchenpos/menugroup/domain/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain.repository;
+package kitchenpos.menugroup.domain.repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
