+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.menugroup.application;
 
 import kitchenpos.common.service.ServiceTest;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.repository.MenuGroupRepository;
+import kitchenpos.menugroup.application.MenuGroupService;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
